+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [6.0.4] - 2026-05-26
+## [Unreleased]
+
+## [6.0.4] - 2026-07-14
 
 ### Fixed
 
@@ -313,4 +315,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#784]: https://github.com/httprb/http/issues/784
 [#785]: https://github.com/httprb/http/issues/785
 [#826]: https://github.com/httprb/http/issues/826
-[unreleased]: https://github.com/httprb/http/compare/v5.3.0...main
+[unreleased]: https://github.com/httprb/http/compare/v6.0.4...HEAD
+[6.0.4]: https://github.com/httprb/http/compare/v6.0.3...v6.0.4
+[6.0.3]: https://github.com/httprb/http/compare/v6.0.2...v6.0.3
+[6.0.2]: https://github.com/httprb/http/compare/v6.0.1...v6.0.2
+[6.0.1]: https://github.com/httprb/http/compare/v6.0.0...v6.0.1
+[6.0.0]: https://github.com/httprb/http/compare/v5.3.1...v6.0.0